requirements.txt
absl-py==1.2.0
actionlib==1.13.2
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
analytics-python==1.4.0
angles==1.9.13
antlr4-python3-runtime==4.9.3
anyio==3.6.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backoff==1.10.0
bcrypt==4.0.0
bondpy==1.8.6
cachetools==5.2.0
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cryptography==38.0.0
cv-bridge==1.16.0
cycler==0.11.0
Cython==0.29.32
cytoolz==0.12.0
dataclasses==0.6
decorator==5.1.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.6
docstring-parser==0.15
dynamic-reconfigure==1.7.3
effdet==0.3.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==1.0.0
fastapi==0.82.0
fastcore==1.3.29
ffmpy==0.3.0
filelock==3.8.0
fire==0.4.0
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
fvcore==0.1.5.post20220512
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.11.0
google-auth-oauthlib==0.4.6
gradio==3.2
graphviz==0.20.1
grpcio==1.48.1
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
icevision==0.12.0
identify==2.5.4
idna==3.3
image-geometry==1.16.0
imageio==2.21.2
importlib-metadata==4.12.0
interactive-markers==1.12.0
iopath==0.1.10
ipykernel==5.5.6
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonargparse==4.9.0
jupyter-client==7.3.5
jupyter-core==4.11.1
kiwisolver==1.4.4
laser-geometry==1.6.7
lightning-bolts==0.5.0
lightning-flash==0.8.0
linkify-it-py==1.0.3
loguru==0.6.0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.0
mdurl==0.1.2
message-filters==1.15.14
monotonic==1.6
multidict==6.0.2
munch==2.5.0
nest-asyncio==1.5.5
networkx==2.8.6
nodeenv==1.7.0
nose==1.3.7
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
orjson==3.8.0
packaging==21.3
pandas==1.4.4
paramiko==2.11.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
portalocker==2.5.1
pre-commit==2.20.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.31
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.4
pycocotools==2.0.4
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pystiche==1.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
python-qt-binding==0.4.4
pytorch-lightning==1.7.4
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.28.1
requests-oauthlib==1.3.1
resnest==0.0.6b20220905
resource-retriever==1.12.7
rfc3986==1.5.0
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rviz==1.14.14
sahi==0.10.5
scikit-image==0.19.3
scipy==1.9.1
seaborn==0.12.0
segmentation-models-pytorch==0.3.0
sensor-msgs==1.13.1
Shapely==1.8.4
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
sniffio==1.3.0
stack-data==0.5.0
starlette==0.19.1
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
tifffile==2022.8.12
timm==0.6.7
toml==0.10.2
toolz==0.12.0
topic-tools==1.15.14
torch==1.10.2
torchmetrics==0.9.3
torchvision==0.11.3
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing-extensions==4.3.0
uc-micro-py==1.0.1
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
virtualenv==20.16.4
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
xacro==1.14.13
yacs==0.1.8
yarl==1.8.1
yolov5-icevision==6.0.0
zipp==3.8.1
